Add first iteration of the web app
This is the first iteration of the web app with placeholder map and scatterplot components.

It uses Bulma for layout and D3 for visualization. At this stage, the data is mostly hardcoded and based on D3 samples. The javascript methods will be refactored in later PRs to take a path to the selected data as a parameter.