Sort and adjust metric titles
Sorted and/or renamed metrics. Fixed 'percent' -> 'percentage' in several places. The overall cancer metric still makes sense to be at the top. We can add a divider or rename it in an upcoming change.

Two metrics were excluded from the selector (oil barrel spills and population over 65). The entries were commented out and the code to render geolocations is still there to support a possible future data source.